requirements.txt
mingus==0.6.1
PyYAML==6.0
six==1.16.0